Refactoring

+++ GameLib/Board.h
@@ -18,7 +18,7 @@ public:
 	~Board();
 
 	const IColumn* GetElement(Position pos) const override;
-	const IColumn* GetElement(int row, int column) const override;
+	const IColumn* GetElement(uint16_t row, uint16_t column) const override;
 	const uint16_t GetSize() const override;
